api/requirements.txt
fastapi
uvicorn[standard]
paho-mqtt
pydantic
